changed file structure of shapes. inertia working on nosecone

+++ src/rocket/misc/finish.cpp
@@ -4,10 +4,9 @@
 
 namespace Rocket{
     // defining constructor
-    Finish::Finish(std::string name, float roughness){
+    Finish::Finish(std::string name, double roughness){
         this->name = name;
         this->roughness = roughness;
-
     }
 
     const Finish regular_paint = Finish("Regular Paint", 60/(std::pow(10,6)));
